refactor: splash_screen, home_screen 분리, MaterialApp
- 화면에 보이는 스크린을 명확하게 구분하여 작성하기 위해, 스플래시 스크린과 홈스크린을 분리했습니다.
  앱 첫 로딩시 runApp -> SplashScreen -> HomeScrren -> OnboardingSreen 순으로 화면이
  이어지는데, OnboardingScreen 이전에 로그인 하도록 로직 수정 예정,
