+++ lib/screen/mypage_screen.dart
@@ -2,9 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:hipspot/component/Favorite/favorite_list.dart';
 import 'package:hipspot/const/font_family.dart';
-import 'package:hipspot/main.dart';
 import 'package:hipspot/api/user.dart';
-import 'package:hipspot/screen/splash_screen.dart';
+import 'package:hipspot/screen/home_screen.dart';
 import 'package:hipspot/utils/authenticate.dart';
 import '../../component/delete_account.dart';
 import '../../const/color/gray_scale_color.dart';
